PostEdit 페이지 컴퍼넌트 수정
